Relocate SMC deploy functions
